refactor(userbodymetrics): add create package to improve organization on interfaces layer
